# [orc] Optimize configuration creating in orc file format