<!-- Please specify the module before the PR name: [core] ... or [flink] ... -->

### Purpose

1. This PR optimizes the creation performance of OrcFileFormat, with an overall performance improvement of at least 500 times. Performance tests can be found in the test files.
2. In actual tests, before optimization, creating an OrcFileFormat took about 6ms, with approximately 7 calls per query, and the scan time was about 40ms+. Under high QPS conditions (50+), the scan time could surge to several hundred milliseconds. After optimization, creating an OrcFileFormat is controlled within about 0.2ms, with approximately 7 calls per query, and the scan time is about 2ms or less.
3. Optimization principle: Each time Hadoop config is created and set, it scans system files.

### Tests

paimon-format/src/test/java/org/apache/paimon/format/orc/OrcFileFormatTest.java

Note: This PR is split from https://github.com/apache/paimon/pull/4497